Changed datatype of `previousBalances` to unit256
The balanceOf mapping returns the value in data type unit256; the `previousBalances` that was summing balanceOf[_from] and balanceOf[_to] and assigning to the unit variable.  Have changed the datatype to unit256.

+++ solidity/token-advanced.sol
@@ -64,7 +64,7 @@ contract TokenERC20 {
         // Check for overflows
         require(balanceOf[_to] + _value > balanceOf[_to]);
         // Save this for an assertion in the future
-        uint previousBalances = balanceOf[_from] + balanceOf[_to];
+        uint256 previousBalances = balanceOf[_from] + balanceOf[_to];
         // Subtract from the sender
         balanceOf[_from] -= _value;
         // Add the same to the recipient
